Rework of XCommands and fixed filter from getting inverted starting value.

+++ TeensyMonitor/Caldera/JSON_Helper.cs
@@ -68,7 +68,7 @@ namespace TeensyMonitor.Caldera
         }
 
         public static string CreateStateChanged(StateChangedMessage message)
-            => CreateStateChanged(message.State);
+            => CreateStateChanged((int)message.State);
 
         private static string CreateStateChanged(int state)
         {
